+++ test/common.py
@@ -210,12 +210,13 @@ def test_scoring(data: Dict[str, Any], translate_params: str, test_similar_score
 
     # Collect scores from output file
     with open(out_path) as score_out:
-        score_scores = [float(line.strip()) for line in score_out]
+        data_scoring = [[float(x) for x in line.strip().split('\t')] for line in score_out]
 
     if test_similar_scores:
-        for inp, translate_json, score_score in zip(data['test_inputs'],
-                                                    data['test_outputs'],
-                                                    score_scores):
+        for inp, translate_json, score_scores in zip(data['test_inputs'],
+                                                     data['test_outputs'],
+                                                     data_scoring):
+            score_score, *factor_scores = score_scores
             translate_tokens = translate_json['translation'].split()
             translate_score = translate_json['score']
             logger.info("tokens: %s || translate score: %.4f || score score: %.4f",
@@ -247,7 +248,7 @@ def _translate_output_is_valid(translate_outputs: List[str]) -> bool:
 def test_odd_even_target_factors(data: Dict):
     num_target_factors = len(data['train_target_factors'])
     for json in data['test_outputs']:
-        factor_keys = [k for k in json.keys() if k.startswith("factor")]
+        factor_keys = [k for k in json.keys() if k.startswith("factor") and not k.endswith("score")]
         assert len(factor_keys) == num_target_factors
         primary_tokens = json['translation'].split()
         secondary_factor_tokens = [json[factor_key].split() for factor_key in factor_keys]