+++ CHANGELOG.md
@@ -11,6 +11,17 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [3.0.2]
+
+### Changed
+
+- `sockeye-translate`: Beam search now computes and returns secondary target factor scores. Secondary target factors 
+  do not participate in beam search, but are greedily chosen at every time step. Accumulated scores for secondary factors
+  are not normalized by length. Factor scores are included in JSON output (``--output-type json``).
+- `sockeye-score` now returns tab-separated scores for each target factor. Users can decide how to combine factor scores
+  depending on the downstream application. Score for the first, primary factor (i.e. output words) are normalized,
+  other factors are not.
+
 ## [3.0.1]
 
 ### Fixed